+++ src/commands/fun/joke.js
@@ -1,5 +1,5 @@
 import { MessageEmbed } from 'discord.js';
-import Command from '../../helpers/Command';
+import Command from '../../structures/Command';
 import { config } from '../../main';
 
 class Joke extends Command {
