+++ src/commands/moderation/purge.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-param-reassign */
-import Command from '../../helpers/Command';
-import { discordError } from '../../helpers/messages';
+import Command from '../../structures/Command';
+import { discordError } from '../../structures/messages';
 import { config, client } from '../../main';
 
 class Purge extends Command {
