+++ src/structures/CreditsManager.js
